feat: use console application
