+++ bin/resque
@@ -21,14 +21,14 @@ define('RESQUE_BIN_DIR', realpath(__DIR__));
 define('RESQUE_DIR', realpath(dirname(RESQUE_BIN_DIR)));
 
 $files = [
-    RESQUE_DIR . '/vendor/autoload.php',
-    RESQUE_DIR . '/../../autoload.php'
+    RESQUE_DIR . '/vendor/autoload.php', // composer dependency
+    RESQUE_DIR . '/../../autoload.php' // stand-alone package
 ];
 
 $loaded = false;
 
 foreach ($files as $file) {
-    if (file_exists($file)) {
+    if (is_file($file)) {
         require_once $file;
         $loaded = true;
         break;
@@ -36,33 +36,13 @@ foreach ($files as $file) {
 }
 
 if (!$loaded) {
-    echo '<pre>You need to set up the project dependencies using the following commands:' . PHP_EOL .
-    "\t" . '$ curl -s http://getcomposer.org/installer | php' . PHP_EOL .
-    "\t" . '$ php composer.phar install</pre>' . PHP_EOL;
-    exit(1);
+    exit(
+        'You need to set up the project dependencies using the following commands:' . PHP_EOL .
+        'curl -sS https://getcomposer.org/installer | php' . PHP_EOL .
+        'php composer.phar install' . PHP_EOL
+    );
 }
 
-$application = new Symfony\Component\Console\Application('php-resque', Resque::VERSION);
-
-$application->add(new Resque\Commands\Clear);
-$application->add(new Resque\Commands\Hosts);
-$application->add(new Resque\Commands\Queues);
-$application->add(new Resque\Commands\Cleanup);
-$application->add(new Resque\Commands\Workers);
-
-$application->add(new Resque\Commands\Job\Queue);
-
-$application->add(new Resque\Commands\Socket\Send);
-$application->add(new Resque\Commands\Socket\Receive);
-$application->add(new Resque\Commands\Socket\Connect);
-
-$application->add(new Resque\Commands\Worker\Start);
-$application->add(new Resque\Commands\Worker\Stop);
-$application->add(new Resque\Commands\Worker\Restart);
-$application->add(new Resque\Commands\Worker\Pause);
-$application->add(new Resque\Commands\Worker\Resume);
-$application->add(new Resque\Commands\Worker\Cancel);
-
-$application->add(new Resque\Commands\SpeedTest);
+$application = new Resque\Console\ResqueApplication();
 
 $application->run();
