+++ src/Resque/Console/Command/Worker/Cancel.php
@@ -9,31 +9,29 @@
  * file that was distributed with this source code.
  */
 
-namespace Resque\Commands\Worker;
+namespace Resque\Command\Worker;
 
 use Resque;
-use Resque\Commands\Command;
+use Resque\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
- * Worker stop command class
+ * Worker cancel command class
  *
  * @author Michael Haynes <mike@mjphaynes.com>
  */
-final class Stop extends Command
+final class Cancel extends Command
 {
     protected function configure(): void
     {
-        $this->setName('worker:stop')
+        $this->setName('worker:cancel')
             ->setDefinition($this->mergeDefinitions([
-                new InputArgument('id', InputArgument::OPTIONAL, 'The id of the worker to stop (optional; if not present stops all workers).'),
-                new InputOption('force', 'f', InputOption::VALUE_NONE, 'Force worker to stop, cancelling any current job.'),
+                new InputArgument('id', InputArgument::OPTIONAL, 'The id of the worker to cancel it\'s running job (optional; if not present cancels all workers).'),
             ]))
-            ->setDescription('Stop a running worker. If no worker id set then stops all workers')
-            ->setHelp('Stop a running worker. If no worker id set then stops all workers');
+            ->setDescription('Cancel job on a running worker. If no worker id set then cancels all workers')
+            ->setHelp('Cancel job on a running worker. If no worker id set then cancels all workers');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
@@ -59,13 +57,18 @@ final class Stop extends Command
             $this->log('<warn>There are no workers on this host</warn>');
         }
 
-        $sig = $input->getOption('force') ? 'TERM' : 'QUIT';
-
         foreach ($workers as $worker) {
-            if (posix_kill($worker->getPid(), constant('SIG'.$sig))) {
-                $this->log('Worker <pop>'.$worker.'</pop> '.$sig.' signal sent.');
+            $packet  = $worker->getPacket();
+            $job_pid = (int)$packet['job_pid'];
+
+            if ($job_pid and posix_kill($job_pid, 0)) {
+                if (posix_kill($job_pid, SIGUSR1)) {
+                    $this->log('Worker <pop>'.$worker.'</pop> running job SIGUSR1 signal sent.');
+                } else {
+                    $this->log('Worker <pop>'.$worker.'</pop> <error>running job SIGUSR1 signal could not be sent.</error>');
+                }
             } else {
-                $this->log('Worker <pop>'.$worker.'</pop> <error>could not send '.$sig.' signal.</error>');
+                $this->log('Worker <pop>'.$worker.'</pop> has no running job.');
             }
         }
