+++ src/Resque/Logger/Processor/ConsoleProcessor.php
@@ -11,7 +11,7 @@
 
 namespace Resque\Logger\Processor;
 
-use Resque\Commands\Command;
+use Resque\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
